Extract file sizes from the Content-Range header
The Content-Range header contains the file size, creating an opportunity
for checking whether the server sent a valid reply: A compliant server
must send a Content-Length that is (file_size - 1) and a Content-Range
header containing "bytes 1-(file_size-1)/file_size".

The background of this commit is that there are servers which have a bug
in their Content-Length calculation, assuming an exclusive range [s,e)
instead of an inclusive one [s,e] in the length calculation. This
results in axel downloading the file one byte short.

The commit resolves this by setting the size to the larger of the two
possible sizes, in the hope that the larger one will always be correct.
An alternative would be to switch off partial downloading entirely in
this case.

See bug #33.