Increase number of max redirects to 20
wget supports up to 20 redirections by default and it makes
sense to have axel behave similarly to avoid mis-expectations
when people move away from wget.

In particular it seems that 5 is a fairly low and gets
hit quite quickly especially when downloading from github
(It happens often when using axel in portage/Gentoo).

Signed-off-by: Antonio Quartulli <a@unstable.cc>