fix use-after-free of axel->conn member
When stopping the download, axel->conn should not
be free'd before having invoked save_state() as it
will be accessed once again to retrieve the current
state.

This use-after-free is easy reproducible on linux when
running axel with at least 10 connections. Probably, due to
some optimization, a smaller number of connections does not
trigger an immediate deallocation of the used heap and
therefore does not exhibit the bug.

Fix this issue by moving the free(axel->conn) instruction
to the end of the cleanup function.

Signed-off-by: Antonio Quartulli <a@unstable.cc>