avoid bad-free when deallocating multiple URLs
When axel is provided with multiple URLs a buffer long
enough to store multiple url_t objects is allocated.

However, the logic assigning each different URL to a
different connection is scrambling the value of axel->url.

This means that upon shutdown the value of axel->url
is not guaranteed to be the one originally returned by malloc()
but may point to the middle. This is likely to trigger
crash during free().

Fix this using a temporary pointer to cycle along the URL
list and leave axel->url as originally assigned.

Signed-off-by: Antonio Quartulli <a@unstable.cc>