make SSL initialization race free
each thread, when started, will attempt to initialize
the SSL component, however, only the first is supposed
to really do that.

A global variable is used to flag is the initialization has
already been performed. However, setting the flag and
performing the initialization must be performed atomically
and protected by a lock so that no other thread can overlap
in this code section.

Signed-off-by: Antonio Quartulli <a@unstable.cc>