Avoid segfault when quiet and alternate behaviour are both specified
When running with arguments "-vvv -q -a" it is possible to observe
a segfault due to axel messing up with the terminal.

In detail: -q closes stdout and redirects it to /dev/null,
but -vvv will turn the verbose conf member to a positive
positive value and thus allow the alternate output (enabled by -a)
to attempt computing the size of the terminal.
This will result in not handled errors and then segfault.

Fix this by forcefully disable alternate output and verbosity
when quiet is specified.

Signed-off-by: Antonio Quartulli <a@unstable.cc>