Check fscanf return value
The fscanf function may fail due to any I/O reading error,
therefore it is better to check its return value and error
out in case of failure.

Note that we can only check for the error value (EOF) and not
for the number of matching pattern because in this case
it may be 0 or 1, based on the config line.

Signed-off-by: Antonio Quartulli <a@unstable.cc>