Compile with -Werror by default
Now that all the warnings are gone, it is recommended
to build with -Werror so that any newly introduced one
will prevent the developer from compiling Axel.

This will hopefully prevent the introduction of new warnings
in our code base.

Consequently, travis-ci will also fail if a new warning is
detected.

Signed-off-by: Antonio Quartulli <a@unstable.cc>