Ensure alternate output can be setup and prevent crash
ioctl() on TIOCGWINSZ might return an error when axel
is not able to access he terminal spec. In this case
the print_alternate_output would just crash because it
does not expect a negative value as terminal width.

Handle this case by disabling the alternate output.

The check against the terminal width is a bit more
relaxed, as it does not make sense to use the alternate
output in a very small window.

Signed-off-by: Antonio Quartulli <a@unstable.cc>