Prevent uninitialized pointer read
We have no guarantee on the value of conn->proxy when
the function is invoked, therefore we have a chance
to dereference puser without having assigned it.

Same can happen for ppass when passed to http_auth_token().

Initialize both at the beginning and perform a NULL check
on puser before accessing it.

Reported by Coverity SCAN

CID: #172884, #172883
Signed-off-by: Antonio Quartulli <a@unstable.cc>