configure: Fix tracking of soft includes
Define an alias for m4_sinclude, so that automake doesn't inject them as
dependencies of the configure script.

The AXEL_VER_READ macro appends the commit ID if it doesn't match the
tag hinted by the specified file.  When working on a Git repository,
this causes configure to be re-generated after each commit.

A minor drawback of fixing this is that when HEAD is updated, the new
commit ID doesn't automatically make it to the VERSION variable. This is
a design limitation of autoconf.

This also removes the need to edit the Makefile.in files for
distribution.

Signed-off-by: Ismael Luceno <ismael@iodev.co.uk>