http: Fix basic auth header generation
Authorization credentials are incorrectly generated depending on
username length.

The algorithm advances 24 bits at a time on it's input, using
chain_next, but it's main loop stops when it finds a pointer to
a zero instead of using the chain_next function, which is wrong.

chain_next takes a NULL-terminated list of pointers to strings,
producing a stream consisting of the concatenation of said strings
followed by an infinite string of zeros.

It is this final string of zeros the loop should be checking for.

Signed-off-by: Shankar <shankar_m@protonmail,com>
[ismael@iodev.co.uk: Edited commit message for normalization and to
 explain the issue further.]
Signed-off-by: Ismael Luceno <ismael@iodev.co.uk>