doc: Fix release date for man-pages on BSD
Process the date at the configure script to handle portability
considerations and simplify the implementation.

The release date is provided in a prettified form to txt2man for
inclusion into the manual pages. This processing is done using the date
command in a non-portable way; BSD and GNU implementations of date are
slightly incompatible.

The GNU implementation of the date command can take an arbitrary date
with the -d option, and the format is automatically detected, but on BSD
implementations it must be invoked like for setting the date but using
the -j option to avoid that step, and the input format must be provided
with the -f option.

Since the -j option isn't supported by GNU, let's exploit it to detect
which variant the system uses.